changed @current_user & @current_login
the currently logged in USER (i.e. the guy in front of the computer) is
now @current_user (formerly @current_login, the current AGENT (i.e. the
entity that buys items and stuff) is now @current_agent (formerly
@current_user)

i might have missed a spot or made a wrong change. if something weird
is displayed on the website, go and check in the source if these two
got confused. (haven't found anything on first sight though)